requirement.txt
defusedxml==0.4.1
dj-database-url==0.3.0
Django==1.7.1
django-appconf==0.6
django-compressor==1.4
djangorestframework==3.0.0
drf-nested-routers==0.9.0
facebook-sdk==0.4.0
gunicorn==19.1.1
oauthlib==0.7.2
Pillow==2.7.0
pip==1.5.6
PyJWT==0.4.0
python-social-auth==0.2.1
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.3.1
setuptools==2.1
six==1.8.0
static3==0.5.1